Refactor CAPU reconfiguration into a transaction object
The previous CAPU configuration methods were somewhat hacky
(eg. CMixer::ExternalSound() called UpdateMixing() with the same
parameters as before, but with different member variables)
and depended on the order of method calls, which was confusing,
and error-prone in theory (but the existing API usages worked fine
in practice, and adding configurable FDS filter would've worked fine).

I replaced the eager update system with CAPUConfig,
a "transaction object" that queues up changes and recomputes the
necessary set of derived data when destroyed.
I find that this approach scales well to more complex scenarios like
exotracker's GUI updates, which proved to be unworkable using
eager updates.