Fix WAV files being opened but never written
If CWavProgressDlg::OnInitDialog() is modified to start a timer with a 0
millisecond period, then sometimes CWavProgressDlg::OnTimer() will run
before the render starts (CSoundGen::OnStartRender() is called). The
dialog calls CSoundGen::IsRendering() which returns false, and then
closes and initiates the next render, skipping the current one
(and leaving the file handle open).

This commit makes CSoundGen::IsRendering() return true before the render
starts, so the dialog properly waits for the render to finish.