Move COM initialization from object to GUI/audio threads
CSoundInterface previously called CoInitializeEx and CoUninitialize in
its constructor and destructor. However this failed on Windows 7: On
startup or after changing configuration, CSoundInterface::OpenChannel
() failed to call IAudioClient::Initialize(), and it instead returned
0x800401f0 CO_E_NOTINITIALIZED. This is because CSoundInterface()
called CoInitializeEx() on the GUI thread, while
CSoundInterface::OpenChannel() is called on the audio thread. This is a
program bug, and Windows 10 and Wine happened to mask this error for
some reason.

To avoid this issue, call CoInitializeEx and CoUninitialize on the
GUI/audio threads' startup/teardown functions (CFamiTrackerApp and
CSoundGen::InitInstance/ExitInstance()), rather than the
CSoundInterface object's constructor and destructor.

Hopefully there aren't problems with accessing COM objects on the wrong
thread. It doesn't matter where IMMDeviceEnumerator/
IMMDeviceCollection/IMMDevice are constructed and accessed, since those
types are only accessed in startup/teardown(which don't need to be
real-time). All real-time logic talks to IAudioClient and
IAudioRenderClient, which are created on the audio thread in
CSoundInterface::OpenChannel() (and hopefully belong to the audio
thread, and if not they use a free-threaded marshaller anyway) and only
accessed on the audio thread.

COM threading is hard. https://github.com/RustAudio/cpal/pull/597,
https://github.com/mozilla/cubeb/issues/534