Move graph data windowing from visualizer thread to audio thread
With this change, the audio thread no longer needs to allocate memory or
acquire locks to send graph data to the visualizer thread.
Unfortunately, the audio thread still needs to lock the document so the
sequencer and driver can read it.

This also eliminates visualizer discontinuities which occur during lag
(the old code treated a double buffer as a queue). Unfortunately the
visualizer is slightly more choppy at buffer sizes 40ms and above on
DirectSound; the visualizer is still smooth on the WASAPI branch.

This change also fixes a race condition causing the visualizer thread's
CVisualizerWnd::ThreadProc() to pass freed memory (pDrawBuffer) to
CVisualizerBase::SetSampleData(). This could occur randomly upon
changing buffer size on DirectSound, or randomly on WASAPI. When it
happens, the visualizer can display corrupted contents or even
segfault.