Move graph data windowing from visualizer thread to audio thread
With this change, the audio thread no longer needs to allocate memory or
acquire locks to send graph data to the visualizer thread.
Unfortunately, the audio thread still needs to lock the document so the
sequencer and driver can read it.

This also eliminates visualizer discontinuities which occur during lag
(the old code treated a double buffer as a queue).