Fix high-res spectrum visualizer at small buffer sizes (#126)
Even with this fix, both the oscilloscope and spectrum visualizer rely
on the visualizer thread processing incoming data faster than the audio
thread can write new data, on a "best-effort" basis. If the visualizer
falls behind, the oscilloscope or spectrum will show a discontinuity in
incoming data. The discontinuity will appear at a random point of the
oscilloscope. In spectrum mode, it will affect the entire spectrum, but
only appears at buffer lengths below 2048 samples.