Fix crash when activating non-initial song then opening document with expansion chip
CFamiTrackerDoc::OpenDocumentNew() and other functions can call
CFamiTrackerDoc::SetupChannels() -> CFamiTrackerDoc::SetMachine
(), during the middle of updating the document. Previously this could
call CDocument::UpdateAllViews() (a method which has no right to
exist), trying to redraw the pattern editor mid-load, when the number
of songs has already been reset to 1 but the active song/track is still
unchanged. If you previously picked a non-initial song,
CPatternEditor::GetFrameCount() calls m_pDocument->GetFrameCount
(GetSelectedTrack()), which indexes out of bounds.

The "solution" is to add a boolean parameter to
CFamiTrackerDoc::SetMachine(), to skip redrawing if the document is in
an inconsistent state. Of course the real fix is to create a
StateTransaction-style type which reloads program state exactly once,
after all changes have been made and the program is in a consistent
state. But FamiTracker is too complex to rewrite state management
entirely.