Fix mostly-theoretical data race when popping from SPSC queues
The reader loses access to the front pointer as soon as it pops from the
queue, since from that point onward, the writer can overwrite it with a
new message. Instead the reader must finish using the pointer or copy
the value out, before popping from the queue.

This bug terrifies me.