Fix debug assert error after failing to export WAV file
We would fail to clear m_pWaveFile if we failed to start rendering. On
the next render attempt, this would cause the assertion
`ASSERT(m_pWaveFile == nullptr)` to fail.

Manually clear m_pWaveFile to prevent this from happening. If mmioOpen
fails, hmmioOut is nullptr and we don't need to call CloseFile().